requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asyncio==3.4.3
asyncpg==0.29.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
customtkinter==5.2.1
darkdetect==0.8.0
ecdsa==0.19.1
fastapi==0.104.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
httplib2==0.31.2
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
numpy==1.26.4
pandas==2.1.3
passlib==1.7.4
proto-plus==1.27.1
protobuf==4.25.8
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
